# docs/swagger.yaml
definitions:
  controllers.PostCreateCartFromItemRequest:
    properties:
      item_id:
        type: integer
      qty:
        type: integer
    required:
    - item_id
    - qty
    type: object
...
  controllers.PostLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controllers.PostRegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - address
    - email
    - name
    - password
    - phone_number
    type: object
  controllers.PutEditCartRequest:
    properties:
      qty:
        type: integer
    required:
    - qty
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a API for Online Shopping Cart API.
  title: Online Shopping Cart API
  version: "1.0"
paths:
  /cart:
    post:
      description: add to own cart from item, need credential
      parameters:
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostCreateCartFromItemRequest'
      produces:
      - application/json
      responses: {}
      summary: add to own cart from item
      tags:
      - Cart
  /cart/{id}:
    delete:
      description: delete item by own cart cart, need credential
      parameters:
      - description: delete by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: delete item by own cart id
      tags:
      - Cart
    get:
      description: get one own carts detail, need credentials
      parameters:
      - description: get detail by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get one own carts detail
      tags:
      - Cart
    put:
      description: update qty of item in cart, delete it if qty 0, failed when qty
        item < qty, need credential
      parameters:
      - description: edit by id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PutEditCartRequest'
      produces:
      - application/json
      responses: {}
      summary: update qty of item in cart
      tags:
      - Cart
  /cart/all:
    get:
      description: get all own carts. need credentials.
      parameters:
      - default: Bearer {token}
        description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: get all own carts
      tags:
      - Cart
...
  /login:
    post:
      description: login account with return JWT token
      parameters:
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostLoginRequest'
      produces:
      - application/json
      responses: {}
      summary: login an account
      tags:
      - Account
...
  /register:
    post:
      description: register account with type CUSTOMER
      parameters:
      - description: Body Request
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/controllers.PostRegisterRequest'
      produces:
      - application/json
      responses: {}
      summary: register an account
      tags:
      - Account
...
swagger: "2.0"